# Improved performance of the version.GetHumanVersion function by 50% on memory allocation.