--- a/.changelog/11507.txt
+++ b/.changelog/11507.txt
@@ -1,0 +1,3 @@
+```release-note:improvement
+version: Improved performance of the version.GetHumanVersion function by 50% on memory allocation.
+```