--- a/version/version_test.go
+++ b/version/version_test.go
@@ -1,0 +1,11 @@
+package version
+
+import (
+	"testing"
+)
+
+func BenchmarkGetHumanVersion(b *testing.B) {
+	for i := 0; i < b.N; i++ {
+		GetHumanVersion()
+	}
+}