Improved performance and added benchmark for func version.GetHumanVersion.

Instructions:
```
$ go install golang.org/x/perf/cmd/benchstat@latest

$ go test -bench=GetHumanVersion -count 5 -benchmem ./version > old.txt
$ go test -bench=GetHumanVersion -count 5 -benchmem ./version > new.txt

$ benchstat old.txt new.txt
```

Results
```
name               old time/op    new time/op    delta
GetHumanVersion-8     177ns ± 2%      80ns ± 3%  -54.53%  (p=0.008 n=5+5)

name               old alloc/op   new alloc/op   delta
GetHumanVersion-8     32.0B ± 0%     16.0B ± 0%  -50.00%  (p=0.008 n=5+5)

name               old allocs/op  new allocs/op  delta
GetHumanVersion-8      3.00 ± 0%      2.00 ± 0%  -33.33%  (p=0.008 n=5+5)
```